Adds skill parse util and dictionary
For efficient user skill storage and searching, skill strings should be normalized. The skillparser.js utility looks for a string matching the user's input in a large map of skills. If no match is found, the input string is returned. If a match is found, the normalized skill is returned. Copied directly out of whobot