Limit Userdb doc projection to just fields reqired for auth
We were previously trying to return the bulk of a user profile document's properties within the Passport Local Strategy authentication callback. This was so that a substantial user object would be returned to clients on login success. This refactor limits the user document fields to just those needed for user authentication. The calling route will handle user object construction after a user has authenticated.