Changes GET all profiles & posts to only return arrays
Both the GET /api/profiles and GET /api/posts routes return an array of objects if the DB call is successful. And both would return a 'No profiles/posts found' message if no documents were found. That's bad form, as it results in an unpredictable API. An endpoint should not return a different data type depending on success/failure of a DB call. This commit changes the 'not found' condition - it now also returns an array (empty in this case).