Refactors getConversations and getConversation handlers
Instead of using mongoose 'populate' methods to add messages to conversations based on arrays of messages stored on conversation documents, we now separately and explicitly search for messages with matching conversation properties. The advantage here is that we no longer need to maintain arrays of message IDs in each conversation record.