Adds 'refresh_token' route and controller to 'apiRoutes'
At some point in this app's history, there was an API enpoint to handle client JWT refreshing. And at some other point in time, that route went away. This commit adds back the route and controller. The route requires a user to have an existing JWT and authenticates them with it. It then generates a new JWT and returns it (and the user's profile) back to the client.